refactor(cli): move the extract group into its own module
Cuts `gpio extract` out of cli/main.py into
geoparquet_io/cli/commands/extract.py, unchanged, along with the
`_deprecated_version_callback` it alone uses and the `ExtractDefaultGroup`
that makes a bare `gpio extract in.parquet out.parquet` mean `extract
geoparquet`.

`tests/test_cli_api_call_parity_scaffold.py` resolved `extract_impl` through
`geoparquet_io.cli.main` and now passes `module=cli_extract`. With that, every
case in the scaffold patches on a group module.

Five pre-existing `arg-type` errors across the arcgis, wfs and carto commands
were suppressed by the mypy per-module override on `geoparquet_io.cli.main`,
so the override entry follows the code rather than the errors being fixed here.

tests/data/cli_surface.json is byte-identical.
